Fix(Core): GM command not working while GM mode is on (#1581)

+++ src/server/scripts/Commands/cs_gobject.cpp
@@ -548,7 +548,7 @@ public:
         stmt->setFloat(5, player->GetPositionY());
         stmt->setFloat(6, player->GetPositionZ());
         stmt->setFloat(7, distance * distance);
-        stmt->setFloat(8, player->GetPhaseMask());
+        stmt->setUInt32(8, player->GetPhaseMask());
         PreparedQueryResult result = WorldDatabase.Query(stmt);
 
         if (result)
